Refactor viewHoroscope()
viewHoroscope() has been renamed to updatePage() and the main 2 responsibilities has been extracted to 2 functions. updatePage() now send the GET request to viewHoroscope.php, and parse the response. The 2 new functions updateButtons() and showSavedHoroscope() get the parsed JSON response in parameter and work with that in contrast to the earlier solution when different functions sent their own GET requests.